Add unit tests
Uses Jasmine and SuperAgent to unit test the function invoker and
HTTP protocol. The bulk of the server code has moved into lib/app.js
in order to make testing easier without needing to bootstrap and
configure a full server process. The app module is a function that
accepts the target function to invoke and returns a configured express
application that is **not** yet listening to a port.

To execute the tests:
- `cd src/main/node`
- `npm install`
- `npm test`