Handle Content-Type headers with a charset
The Content-Type header can also contain a charset param.

```http
Content-Type: text/plain; charset=utf-8
```

Previously this would cause the gRPC handler to choke. Now the charset
is ignored. In the future we should correctly parse the charset and
decode the payload accordingly. For now, all text content is assumed to
be utf-8.

Fixes: #29