Allow functions to receive messages or payloads
Currently, functions only recieve message payloads when the function is
invoked. This is akin to an HTTP server that can only the the entity of
a request, or worse, can only set the entity of a response. Each message
in riff contains headers in addition to the payload.

Now functions can opt-in to receive messages or headers. Payloads are
still available by default.

```js
const { Message } = require('@projectriff/message');

Message.install();

module.exports = message => {
    // validate GitHub WebHook signature
    const signature = message.headers.getValue('X-Hub-Signature');
    ...
};
module.exports.$argumentType = 'message';
```

The [@projectriff/message](https://github.com/projectriff/node-message)
repository defines the Message and Headers types that are used. They:

- convert to and from the raw format defined by
  [function-proto](https://github.com/projectriff/riff/tree/master/function-proto)
- treat header names as case insensitive, while also case preserving
- handle multiple header values
- are immutable
- use a fluent builder pattern to create messages

The npm package can be installed and required by your function to
produce a message with custom headers. Any function that emits a value
that is an instance of Message will be treated as a message, otherwise,
it will be treated as a payload.

```
const { Message } = require('@projectriff/message');

let count = 0;
let instance = Math.round(Math.random() * 10000)

module.exports = name => {
    return Message.builder()
        .addHeader('X-Instance-ID', instance)
        .addHeader('X-Instance-Count', count++)
        .payload(`Hello ${name}!`)
        .build();
};
```

Issue: #41