Restore HTTP protocol support
and a significant internal refactor to make supporting new protocols
easier by decoupling the protocol from the interaction model and
argument type. The interaction model and argument type transformer are
now injected into the protocol handler.

All interaction models are now fundementally based on node streams. The
HTTP interaction model creates a bounded input stream with a single
message and expected the output stream to produce a single message as
the response. The request-reply interation model bridges from a stream
to a single function invocation.