Fix #130,#131. Improve pipeline error handling and related tests
 1. prepend test error listeners (see explanation below)
 2. observe 'close' events after 'error', not 'finish' (source:
https://github.com/nodejs/help/issues/2412)
 3. make sure pipeline is ended by error handling and not by
propagation of the piped source regular completion ({end: false})
 4. create pipeline stream with autodestroy enabled, so that the
stream automatically closes upon error
 5. reorder and document streaming pipeline class for clarity

When an marshaller's or unmarshaller's error happens, the streaming
pipeline intercepts (`unmarshaller.on('error', ...)`) and emits it
itself (`this.emit('error', ...)`), cleans up input/output streams and
completes (which means: emits a `finish` event).

Given `emit` is completely synchronous, this means that the pipeline's
`finish` listeners will always be executed as part of the synchronous
execution of the pipeline's own `error` handler (which is, by default,
the first pipeline's `error` handler to be executed as it is
registered in the pipeline's constructor).

The error case tests assumed that the test `error` handler will always
be executed before the test `finish` handler.  This can only work if the
test `error` handler is prepended to the list of `error` handler.