Change output content type preference to favor json
Currently, output that uses a wildcard accept header defaults to plain
text. While this is the most compatible for clients, it returns garbage
when the function returns an object, or something with structure.

JSON, while less understood by clients, is able to handle a wider
variety of function outputs. Ultimately, clients should send an Accept
header to pick the type they actually want.

Refs #180